Refactor syntax generators into SyntaxGenerators class
